# connectin-api-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Connect'in API",
        "description": "Documentation Swagger de l'API Connect'in",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Afficher le profil connecté",
                "operationId": "24387fa9f986421b9a77da1444399e3d",
                "responses": {
                    "200": {
                        "description": "Profil"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Mettre à jour le profil",
                "operationId": "252f2e8313e0796f97867101fe16c080",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Ada",
                                        "nullable": true
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Lovelace",
                                        "nullable": true
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Ma nouvelle bio",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ada@connectin.com",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Supprimer le compte",
                "operationId": "7da9edcac4557de03de81c83a8e56feb",
                "parameters": [
                    {
                        "name": "delete_all",
                        "in": "query",
                        "description": "Option : Si le front envoie ?delete_all=true, on supprime le contenu associé (posts + commentaires). Sinon on garde le contenu et le nom/prénom devient \"Utilisateur supprimé\".",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "MotDePasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte supprimé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Mot de passe incorrect / validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/posts": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Lister les posts du profil connecté",
                "operationId": "d78937e7480e128c425b4c8feab68eb4",
                "responses": {
                    "200": {
                        "description": "Liste des posts du user connecté"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Voir le profil d'un utilisateur",
                "operationId": "efd15249fdc88d1ad733683efdf695f3",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID du user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil user"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "User introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{user}/posts": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Lister les posts d'un utilisateur",
                "operationId": "4b0345bc931da0f1edc81f9efcb8894f",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID du user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des posts du user"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "User introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/picture": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Mettre à jour la photo de profil",
                "operationId": "ec3c6d05a6bb83571c2df581c5324776",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "profile_picture"
                                ],
                                "properties": {
                                    "profile_picture": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo mise à jour"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Supprimer la photo de profil",
                "operationId": "f3c5e771778dc3409b7890680985864d",
                "responses": {
                    "200": {
                        "description": "Photo de profil supprimée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrer le token sous la forme: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}